feat: 아티클 카테고리 조회 필터 적용
